support floating point source dimensions

+++ src/main/kotlin/com/painkiller/layouts/Inset.kt
@@ -3,7 +3,7 @@ package com.painkiller.layouts
 data class Inset(
   val width: Int,
   val height: Int,
-  val sourceWidth: Int,
-  val sourceHeight: Int,
+  val sourceWidth: Double,
+  val sourceHeight: Double,
   val margin: Int,
 )
