+++ src/main/kotlin/com/painkiller/layouts/InsetService.kt
@@ -17,14 +17,14 @@ class InsetService {
   }
 
   private fun marginLeft(inset: Inset): Int {
-    if (inset.sourceHeight == 0) return 0
+    if (inset.sourceHeight == 0.0) return 0
     val innerWidth = inset.width - inset.margin * 2
-    return (inset.width - innerWidth * inset.sourceWidth / inset.sourceHeight) / 2
+    return (inset.width - (innerWidth * inset.sourceWidth / inset.sourceHeight).toInt()) / 2
   }
 
   private fun marginTop(inset: Inset): Int {
-    if (inset.sourceWidth == 0) return 0
+    if (inset.sourceWidth == 0.0) return 0
     val innerHeight = inset.height - inset.margin * 2
-    return (inset.height - innerHeight * inset.sourceHeight / inset.sourceWidth) / 2
+    return (inset.height - (innerHeight * inset.sourceHeight / inset.sourceWidth).toInt()) / 2
   }
 }
